unify further mazerunner scenarios
