+++ mazerunner-scenarios/src/main/java/com/bugsnag/android/mazerunner/scenarios/AppVersionScenario.kt
@@ -1,19 +1,22 @@
 package com.bugsnag.android.mazerunner.scenarios
 
 import android.content.Context
+
 import com.bugsnag.android.Bugsnag
 import com.bugsnag.android.Configuration
 
 /**
- * Sends a handled exception to Bugsnag, which contains metadata that should be filtered
+ * Sends a handled exception to Bugsnag, which overrides the app version
  */
-internal class AutoFilterScenario(config: Configuration,
+internal class AppVersionScenario(config: Configuration,
                                   context: Context) : Scenario(config, context) {
+    init {
+        config.setAutoCaptureSessions(false)
+        config.setAppVersion("1.2.3.abc")
+    }
+
     override fun run() {
         super.run()
-        Bugsnag.addToTab("user", "password", "hunter2")
-        Bugsnag.addToTab("custom", "password", "hunter2")
-        Bugsnag.addToTab("custom", "foo", "hunter2")
         Bugsnag.notify(generateException())
     }
