+++ mazerunner-scenarios/src/main/java/com/bugsnag/android/mazerunner/scenarios/EmptyStacktraceScenario.kt
