+++ mazerunner-scenarios/src/main/java/com/bugsnag/android/mazerunner/scenarios/StoppedSessionScenario.kt
@@ -5,13 +5,15 @@ import com.bugsnag.android.Bugsnag
 import com.bugsnag.android.Configuration
 import android.os.Handler
 import android.os.HandlerThread
-import com.bugsnag.android.flushAllSessions
 
 /**
  * Sends an exception after stopping the session
  */
-internal class NewSessionScenario(config: Configuration,
+internal class StoppedSessionScenario(config: Configuration,
                                       context: Context) : Scenario(config, context) {
+    init {
+        config.setAutoCaptureSessions(false)
+    }
 
     override fun run() {
         super.run()
@@ -19,19 +21,14 @@ internal class NewSessionScenario(config: Configuration,
         val thread = HandlerThread("HandlerThread")
         thread.start()
 
-        Handler(thread.looper).post {
+        Handler(thread.looper).post(Runnable {
             // send 1st exception which should include session info
             client.startSession()
-            flushAllSessions()
             client.notifyBlocking(generateException())
 
-            // stop tracking the existing session
+            // send 2nd exception which should not include session info
             client.stopSession()
-
-            // send 2nd exception which should contain new session info
-            client.startSession()
-            flushAllSessions()
             client.notifyBlocking(generateException())
-        }
+        })
     }
 }
