--- tests/features/fixtures/mazerunner/src/main/java/com/bugsnag/android/mazerunner/scenarios/AppVersionScenario.kt
@@ -1,22 +1,0 @@
-package com.bugsnag.android.mazerunner.scenarios
-
-import android.content.Context
-import com.bugsnag.android.Bugsnag
-import com.bugsnag.android.Configuration
-
-/**
- * Sends a handled exception to Bugsnag, which overrides the app version
- */
-internal class AppVersionScenario(config: Configuration,
-                                  context: Context) : Scenario(config, context) {
-
-    init {
-        config.setAppVersion("1.2.3.abc")
-    }
-
-    override fun run() {
-        super.run()
-        Bugsnag.notify(generateException())
-    }
-
-}
