--- tests/features/fixtures/mazerunner/src/main/java/com/bugsnag/android/mazerunner/scenarios/IgnoredExceptionScenario.kt
@@ -1,22 +1,0 @@
-package com.bugsnag.android.mazerunner.scenarios
-
-import android.content.Context
-import com.bugsnag.android.Bugsnag
-import com.bugsnag.android.Configuration
-
-/**
- * Attempts to send an ignored handled exception to Bugsnag, which should not result
- * in any operation.
- */
-internal class IgnoredExceptionScenario(config: Configuration,
-                                        context: Context) : Scenario(config, context) {
-    init {
-        config.setIgnoreClasses(listOf("java.lang.RuntimeException"))
-    }
-
-    override fun run() {
-        super.run()
-        throw RuntimeException("Should never appear")
-    }
-
-}
