--- tests/features/fixtures/mazerunner/src/main/java/com/bugsnag/android/mazerunner/scenarios/MinimalUnhandledExceptionScenario.kt
@@ -1,22 +1,0 @@
-package com.bugsnag.android.mazerunner.scenarios
-
-import android.content.Context
-import com.bugsnag.android.Configuration
-import java.io.File
-
-/**
- * Sends an unhandled exception to Bugsnag.
- */
-internal class MinimalUnhandledExceptionScenario(config: Configuration,
-                                                 context: Context) : Scenario(config, context) {
-    init {
-        config.setAutoCaptureSessions(false)
-        disableAllDelivery(config)
-    }
-
-    override fun run() {
-        super.run()
-        throw java.lang.IllegalStateException("Whoops")
-    }
-
-}
